feat/refactor: altered file structure setup and compartmentalised functions a little more
